requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.26.0
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
alembic==1.13.1
asyncpg==0.29.0
stripe==7.0.0
gunicorn
